+++ src/MAUI/Maui.Samples/Samples/UtilityNetwork/SnapGeometryEditsWithUtilityNetworkRules/SnapGeometryEditsWithUtilityNetworkRules.xaml.cs
@@ -1,10 +1,10 @@
-// Copyright 2025 Esri.
+﻿// Copyright 2025 Esri.
 //
 // Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at: http://www.apache.org/licenses/LICENSE-2.0
 //
-// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an 
-// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific 
+// Unless required by applicable law or agreed to in writing, software distributed under the License is distributed on an
+// "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the specific
 // language governing permissions and limitations under the License.
 
 using Esri.ArcGISRuntime.Data;
@@ -17,22 +17,22 @@ using Esri.ArcGISRuntime.UI;
 using Esri.ArcGISRuntime.ArcGISServices;
 using Esri.ArcGISRuntime.UI.Controls;
 using System;
+using System.Collections.Generic;
+using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
-using Microsoft.UI.Xaml;
-using Microsoft.UI.Xaml.Input;
 
-namespace ArcGIS.WinUI.Samples.SnapGeometryWithRules
+namespace ArcGIS.Samples.SnapGeometryEditsWithUtilityNetworkRules
 {
     [ArcGIS.Samples.Shared.Attributes.Sample(
-        "Snap geometry with rules",
-        "Geometry",
-        "Perform snap geometry edits with rules",
-        "")]
+        name: "Basic readme template",
+        category: "Utility network",
+        description: "",
+        instructions: "")]
     [ArcGIS.Samples.Shared.Attributes.OfflineData()]
-    public partial class SnapGeometryWithRules
+    public partial class SnapGeometryEditsWithUtilityNetworkRules
     {
-        public SnapGeometryWithRules()
+        public SnapGeometryEditsWithUtilityNetworkRules()
         {
             InitializeComponent();
             _ = Initialize();
