+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/Analysis/DistanceMeasurement/readme.metadata.json
@@ -14,10 +14,7 @@
     ],
     "offline_data": [],
     "redirect_from": [
-        "/net/latest/wpf/sample-code/configure-electronic-navigational-charts.htm",
-        "net/latest/wpf/sample-code/select-enc-features.htm",
-        "net/latest/wpf/sample-code/change-enc-display-settings.htm",
-        "net/latest/wpf/sample-code/add-enc-exchange-set.htm"
+        "/net/latest/winui/sample-code/distance-measurement-analysis.htm"
     ],
     "relevant_apis": [
         "AnalysisOverlay",
