--- src/WinUI/ArcGIS.WinUI.Viewer/Samples/Geometry/SnapGeometryWithRules/readme.metadata.json
@@ -1,19 +1,0 @@
-{
-    "category": "Geometry",
-    "description": "Perform snap geometry edits with rules",
-    "formal_name": "SnapGeometryWithRules",
-    "ignore": false,
-    "images": [
-        "SnapGeometryWithRules.jpg"
-    ],
-    "keywords": [
-    ],
-    "redirect_from": [
-        "/net/latest/wpf/sample-code/SnapGeometryWithRules.htm"
-    ],
-    "relevant_apis": [
-    ],
-    "snippets": [
-    ],
-    "title": "Snap geometry with rules"
-}
