+++ src/WinUI/ArcGIS.WinUI.Viewer/Samples/UtilityNetwork/SnapGeometryEditsWithUtilityNetworkRules/SnapGeometryEditsWithUtilityNetworkRules.xaml
@@ -1,5 +1,5 @@
 <UserControl 
-    x:Class="ArcGIS.WinUI.Samples.SnapGeometryWithRules.SnapGeometryWithRules"
+    x:Class="ArcGIS.WinUI.Samples.SnapGeometryEditsWithUtilityNetworkRules.SnapGeometryEditsWithUtilityNetworkRules"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:esri="using:Esri.ArcGISRuntime"
